split message definition parsing out from MessageReader
Reading a message definition and building a de-serializer are two
separate steps. Splitting the definition parsing from de-serialization
decouples the two and allows for refactoring each while still
maintaining interface boundaries.
