+++ lib/MessageReader.js
@@ -6,6 +6,7 @@
 
 import int53 from "int53";
 import { extractTime } from "./fields";
+import { parseMessageDefinition } from "./parseMessageDefinition";
 
 // this has hard-coded buffer reading functions for each
 // of the standard message types http://docs.ros.org/api/std_msgs/html/index-msg.html
@@ -116,94 +117,6 @@ class StandardTypeReader {
   }
 }
 
-function normalizeType(type) {
-  // Normalize deprecated aliases.
-  let normalizedType = type;
-  if (type === "char") normalizedType = "uint8";
-  if (type === "byte") normalizedType = "int8";
-  return normalizedType;
-}
-
-// represents a single line in a message definition type
-// e.g. 'string name' 'CustomType[] foo' 'string[3] names'
-function newDefinition(type, name, isArray = false, arrayLength = undefined) {
-  const normalizedType = normalizeType(type);
-  return {
-    type: normalizedType,
-    name,
-    isArray,
-    arrayLength,
-    isComplex: !StandardTypeReader.prototype[normalizedType],
-  };
-}
-
-// represents a definition of a custom type in a message definition
-function newComplexType(name, definitions) {
-  return { name, definitions };
-}
-
-function newCustomType(type, name) {
-  return { isCustom: true, name: name, type: type };
-}
-
-const buildType = (lines, customParsers) => {
-  const instructions = [];
-  let complexTypeName;
-  let customType;
-  lines.forEach((line) => {
-    // remove comments and extra whitespace from each line
-    const splits = line
-      .replace(/#.*/gi, "")
-      .split(" ")
-      .filter((word) => word);
-    if (!splits[1]) {
-      return;
-    }
-    // consume comments
-    const type = splits[0].trim();
-    const name = splits[1].trim();
-    if (Object.keys(customParsers).indexOf(type) > -1) {
-      customType = newCustomType(type, name);
-    }
-    if (type === "MSG:") {
-      complexTypeName = name;
-    } else if (name.indexOf("=") > -1 || splits.indexOf("=") > -1) {
-      // constant type parsing
-      const matches = line.match(/(\S+)\s*=\s*(.*)\s*/);
-      let value = matches[2];
-      if (type !== "string") {
-        try {
-          value = JSON.parse(value.replace(/\s*#.*/g, ""));
-        } catch (error) {
-          // eslint-disable-next-line no-console
-          console.warn(`Error in this constant definition: ${line}`);
-          throw error;
-        }
-        if (type === "bool") value = Boolean(value);
-      }
-      if ((type.includes("int") && value > Number.MAX_SAFE_INTEGER) || value < Number.MIN_SAFE_INTEGER) {
-        // eslint-disable-next-line no-console
-        console.warn(`Found integer constant outside safe integer range: ${line}`);
-      }
-      instructions.push({
-        type: normalizeType(type),
-        name: matches[1],
-        isConstant: true,
-        value,
-      });
-    } else if (type.indexOf("]") === type.length - 1) {
-      // array type parsing
-      const typeSplits = type.split("[");
-      const baseType = typeSplits[0];
-      const arrayLength = parseInt(typeSplits[1].replace("]", ""), 10) || undefined;
-      instructions.push(newDefinition(baseType, name, true, arrayLength));
-    } else {
-      instructions.push(newDefinition(type, name));
-    }
-  });
-  return newComplexType(complexTypeName, customType || instructions);
-};
-
 const findTypeByName = (types, name = "") => {
   const matches = types.filter((type) => {
     const typeName = type.name || "";
@@ -223,9 +136,6 @@ const findTypeByName = (types, name = "") => {
 };
 
 const constructorBody = (type) => {
-  if (type.definitions.isCustom) {
-    return ";";
-  }
   return type.definitions
     .map((def) => {
       return `this.${def.name} = undefined`;
@@ -260,11 +170,6 @@ Record.${friendlyName(t.name)} = function() {
   let stack = 0;
   const getReaderLines = (type, fieldName = "record") => {
     let readerLines = [];
-    if (type.definitions.isCustom) {
-      const def = type.definitions;
-      readerLines.push(`${fieldName}.${def.name} = customParsers["${def.type}"](reader)`);
-      return readerLines;
-    }
     type.definitions.forEach((def) => {
       if (def.isConstant) {
         return;
@@ -329,7 +234,7 @@ Record.${friendlyName(t.name)} = function() {
 
   const lines = getReaderLines(unnamedType).join("\n");
   const readerFn = `
-  return function read(reader, customParsers) {
+  return function read(reader) {
     var record = new Record();
     ${lines}
     return record;
@@ -345,60 +250,22 @@ Record.${friendlyName(t.name)} = function() {
     throw e;
   }
 
-  return function(buffer, customParsers) {
+  return function(buffer) {
     const reader = new StandardTypeReader(buffer);
-    return _read(reader, customParsers);
+    return _read(reader);
   };
 };
-
-export function getTypes(messageDefinition = "", customParsers = {}) {
-  // read all the lines and remove empties
-  const allLines = messageDefinition
-    .split("\n")
-    .map((line) => line.trim())
-    .filter((line) => line);
-
-  let definitionLines = [];
-  const types = [];
-  // group lines into individual definitions
-  allLines.forEach((line) => {
-    // skip comment lines
-    if (line.indexOf("#") === 0) {
-      return;
-    }
-    // definitions are split by equal signs
-    if (line.indexOf("==") === 0) {
-      types.push(buildType(definitionLines, customParsers));
-      definitionLines = [];
-    } else {
-      definitionLines.push(line);
-    }
-  });
-  types.push(buildType(definitionLines, customParsers));
-
-  // Fix up complex type names
-  types.forEach(({ definitions }) => {
-    if (definitions.isCustom) return;
-    definitions.forEach((definition) => {
-      if (definition.isComplex) {
-        definition.type = findTypeByName(types, definition.type).name;
-      }
-    });
-  });
-
-  return types;
-}
 
 export default class MessageReader {
   // takes a multi-line string message definition and returns
   // a message reader which can be used to read messages based
   // on the message definition
-  constructor(messageDefinition = "", customParsers = {}) {
-    this.customParsers = customParsers;
-    this.reader = createParser(getTypes(messageDefinition, customParsers));
+  constructor(messageDefinition = "") {
+    const definitions = parseMessageDefinition(messageDefinition);
+    this.reader = createParser(definitions);
   }
 
   readMessage(buffer) {
-    return this.reader(buffer, this.customParsers);
+    return this.reader(buffer);
   }
 }
