+++ lib/index.js
@@ -7,12 +7,13 @@
 import Bag from "./bag";
 
 import BagReader from "./BagReader";
-import MessageReader, { getTypes } from "./MessageReader";
+import MessageReader from "./MessageReader";
+import { parseMessageDefinition as getTypes, builtins as rosPrimitiveTypes } from "./parseMessageDefinition";
 import Time from "./Time";
 
 // export this as a named export for es5 module compatibility
 const open = Bag.open;
 
-export { Time, BagReader, MessageReader, open, getTypes };
+export { Time, BagReader, MessageReader, open, getTypes, rosPrimitiveTypes };
 
 export default Bag;
